// src/models/IDocument.ts
export interface IDocument {
  Account_First_Name?: string;
  Account_Last_Name?: string;
  Account_User_Name?: string;
  Account_Email?: string;
  // Account_Email_OTP?: string;
  Account_Email_ForgetPass?: string;
  Account_Password?: string;
  Account_Password_newPassword?: string;
  Account_OTP?: string;
  Account_UserData_Store?: any;
  Account_allowsNewsletter?: any;
  Account_Terms_And_Condition?: any;
}